requirements.txt
groq
discord
python-dotenv
pynacl
yt-dlp
ffmpeg
